рефакторинг утилит, разбивка на части
