+++ src/utils/utils-common.js
@@ -1,0 +1,18 @@
+const draft = ``;
+const HeadingTitle = {
+  LIST: `Trip events`,
+  MENU: `Switch trip view`,
+  FILTER: `Filter events`,
+};
+
+const getRadomNum = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+export {
+  draft,
+  getRadomNum,
+  HeadingTitle,
+};
