+++ src/utils/utils-render.js
@@ -1,15 +1,8 @@
-import dayjs from 'dayjs';
-import AbstractView from './view/absract.js';
+import AbstractView from '../view/absract.js';
 
-const draft = ``;
 const RenderPosition = {
   AFTERBEGIN: `afterbegin`,
   BEFOREEND: `beforeend`
-};
-const HeadingTitle = {
-  LIST: `Trip events`,
-  MENU: `Switch trip view`,
-  FILTER: `Filter events`,
 };
 
 const render = (container, child, position) => {
@@ -53,45 +46,10 @@ const createElement = (template) => {
   return container.firstChild;
 };
 
-const getRadomNum = (min, max) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-const generateRandomIndex = (data) => {
-  const randomData = data[getRadomNum(0, data.length - 1)];
-  return randomData;
-};
-
-const humanizeDate = (format, date) => {
-  return dayjs(date).format(format);
-};
-
-const checkEmptyData = (data) => data.length === 0;
-
-const getEmptyDataClassName = (data) => checkEmptyData(data) ? `visually-hidden` : ``;
-
-const filterData = (data, parameter) => {
-  const sortedEvents = data.slice();
-  sortedEvents.sort((left, right) => {
-    return left[parameter] - right[parameter];
-  });
-  return sortedEvents;
-};
-
 export {
-  draft,
   RenderPosition,
   render,
   renderTemplate,
   createElement,
-  getRadomNum,
-  generateRandomIndex,
-  humanizeDate,
-  checkEmptyData,
-  getEmptyDataClassName,
-  filterData,
-  HeadingTitle,
   replace,
 };
