+++ src/view/new-event.js
@@ -1,6 +1,7 @@
-import {humanizeDate, getEmptyDataClassName, checkEmptyData, draft} from '../utils.js';
+import {humanizeDate, getEmptyDataClassName, checkEmptyData} from '../utils/utils-event.js';
+import {draft} from '../utils/utils-common.js';
 import {destinations} from '../mock/mock-event.js';
-import AbstractView from './absract.js';
+import {AbstractView} from './absract.js';
 
 const createNewPointTemplate = (event) => {
   const {
